+++ apps/ledger-live-mobile/src/const/navigation.ts
@@ -505,6 +505,9 @@ export enum ScreenName {
   Stake = "Stake",
 
   RedirectToOnboardingRecoverFlow = "RedirectToOnboardingRecoverFlow",
+
+  AnalyticsOptInPromptMain = "AnalyticsOptInPromptMain",
+  AnalyticsOptInPromptDetails = "AnalyticsOptInPromptDetails",
 }
 
 export enum NavigatorName {
@@ -615,4 +618,6 @@ export enum NavigatorName {
 
   PostOnboarding = "PostOnboarding",
   SyncOnboarding = "SyncOnboarding",
+
+  AnalyticsOptInPrompt = "AnalyticsOptInPrompt",
 }