+++ apps/ledger-live-mobile/src/screens/AnalyticsOptInPrompt/Common.tsx
@@ -1,0 +1,28 @@
+import { Flex } from "@ledgerhq/native-ui";
+import styled from "styled-components/native";
+
+export const View = styled(Flex).attrs({
+  width: "100%",
+})``;
+
+export const Container = styled(Flex).attrs({
+  paddingHorizontal: 16,
+  flex: 1,
+  justifyContent: "space-between",
+})``;
+
+export const Titles = styled(Flex).attrs({
+  pt: 6,
+  width: "100%",
+})``;
+
+export const Content = styled(Flex).attrs({
+  pt: 6,
+  width: "100%",
+})``;
+
+export const Bottom = styled(Flex).attrs({
+  paddingBottom: 7,
+  paddingTop: 0,
+  width: "100%",
+})``;